feat: add composable for handling tabs
- add useTabs composable
- use useTabs in all instances where tabs management is needed
